api/requirements.txt
Flask==2.2.2
Flask_Cors==3.0.10
Flask_MySQLdb==1.0.1
googletrans==3.1.0a0
nltk==3.7
protobuf==4.21.12
